pyproject.toml
[tool.poetry]
name = "user-auth"
version = "0.1.0"
description = ""
authors = ["M NOMAN CH <noman.chaudhary@tylextech.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.30"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
